# api/openapi.yaml
openapi: 3.0.3
info:
  title: PGDock API
  version: 1.0.0
  description: |
    Control-plane API for PGDock. This file is the single source of truth:
    Go server interfaces (oapi-codegen) and TypeScript client types
    (openapi-typescript) are generated from it with `make generate`.
servers:
  - url: /
paths:
...
  /api/v1/auth/login:
    post:
      operationId: postAuthLogin
      summary: Email + password login; returns a TOTP challenge
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Password accepted; complete login with /auth/totp.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginChallenge"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/auth/totp:
    post:
      operationId: postAuthTotp
      summary: Complete login with a TOTP code
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpRequest"
      responses:
        "200":
          description: Logged in; the session cookie is set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionState"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/auth/reauth:
    post:
      operationId: postAuthReauth
      summary: Step-up authentication for destructive actions
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReauthRequest"
      responses:
        "204":
          description: Step-up authentication granted.
        default:
          $ref: "#/components/responses/Error"
  /api/v1/auth/logout:
    post:
      operationId: postAuthLogout
      summary: End the current session
      tags: [auth]
      responses:
        "204":
          description: Logged out.
        default:
          $ref: "#/components/responses/Error"
  /api/v1/me:
    get:
      operationId: getMe
      summary: The signed-in user
      tags: [account]
      responses:
        "200":
          description: The signed-in user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateMe
      summary: Change your profile
      tags: [account]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/setup/begin:
    post:
      operationId: beginSetup
      summary: First-run setup, step 1 - owner account and TOTP enrolment
      description: |
        Only before any operator exists. `setup_code` is printed in the
        pgdock-server log. Returns the TOTP secret to enrol; nothing is
        created until `/setup/complete` proves a code.
      tags: [setup]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupBeginRequest"
      responses:
        "200":
          description: TOTP enrolment details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupEnrollment"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/setup/complete:
    post:
      operationId: completeSetup
      summary: First-run setup, step 2 - confirm TOTP and create the owner
      tags: [setup]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupCompleteRequest"
      responses:
        "200":
          description: Owner created and signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionState"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/settings/db-host:
    put:
      operationId: putDbHost
      summary: Set the hostname clients use for databases
      description: Also starts obtaining the pooler's TLS certificate for it.
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DbHostRequest"
      responses:
        "200":
          description: Updated settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettings"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/settings/db-host/check:
    post:
      operationId: checkDbHost
      summary: Check that a hostname resolves to this server
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DbHostRequest"
      responses:
        "200":
          description: DNS check result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DnsCheck"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/dev/operations:
    post:
      operationId: createDevOperation
      summary: Enqueue a dummy operation (development only)
      description: |
        Available only when the server runs with PGDOCK_DEV_ENDPOINTS=true;
        otherwise returns 404. Enqueues a `noop` operation for exercising
        the queue end to end.
      tags: [dev]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoopParams"
      responses:
        "202":
          description: The operation was queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects:
    get:
      operationId: listProjects
      summary: List projects you can see in an organisation, newest first
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/OrgQuery"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProjectStatus"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Live projects.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createProject
      summary: Create a shared-tier project
      description: |
        Records the project and queues a `create` operation. The response
        carries the password and full connection URLs exactly once; PGDock
        stores only the SCRAM verifier. The URLs work once the operation
        succeeds (stream it to follow progress).
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "202":
          description: Project recorded and create operation queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCredentials"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}:
    get:
      operationId: getProject
      summary: Project detail and connection info (no password)
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: The project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteProject
      summary: Delete a project
      description: |
        `confirm` must equal the project's name, and the session must have
        re-authenticated recently (`POST /auth/reauth`); otherwise 403 with
        code `reauth_required`. Queues a `delete` operation that takes a
        final backup (kept 30 days; `skip_final_backup` skips it), removes
        the pooler route, and drops the database and role.
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: confirm
          in: query
          required: true
          schema:
            type: string
        - name: skip_final_backup
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "202":
          description: Delete operation queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}/settings:
    patch:
      operationId: updateProject
      summary: Update a project's name, description, or guardrails
      description: |
        Changes to guardrails that live on the backend or the pooler
        (connection limit, timeouts, pool size) queue an `apply_settings`
        operation, returned alongside the project.
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: The updated project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectUpdated"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}/rotate-password:
    post:
      operationId: rotateProjectPassword
      summary: Rotate the project password
      description: |
        Returns the new password once and queues a `rotate` operation. The
        new password works, and the old one stops working, when it succeeds.
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "202":
          description: Rotate operation queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCredentials"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}/backups:
    post:
      operationId: createProjectBackup
      summary: Back up a project now
      description: Queues a `backup` operation (a logical backup, kept by the nightly retention policy).
      tags: [backups]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "202":
          description: Backup operation queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/backups/{id}/restore:
    post:
      operationId: restoreBackup
      summary: Restore a backup into a new project or in place
      description: |
        `mode: new` (the default, non-destructive) creates a project named
        `name` and returns its credentials once. `mode: in_place` replaces the
        backup's own project: `confirm` must equal the project's name and the
        session must have re-authenticated recently (403 `reauth_required`).
        It takes a safety backup first.
      tags: [backups]
      parameters:
        - $ref: "#/components/parameters/BackupID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreRequest"
      responses:
        "202":
          description: Restore operation queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/restore-tests:
    post:
      operationId: runRestoreTest
      summary: Run the restore test now
      description: Restores the latest backup of a random project (or `project_id`) into a scratch database, counts every table, and drops it.
      tags: [backups]
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "202":
          description: Restore test queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/settings/storage:
    get:
      operationId: getStorageSettings
      summary: The backup storage target (never the secret key)
      tags: [settings]
      responses:
        "200":
          description: Storage settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageSettings"
        default:
          $ref: "#/components/responses/Error"
    put:
      operationId: putStorageSettings
      summary: Save the backup storage target after a live test
      description: Runs the write/read/delete test and saves only if it passes. An empty `secret_key` keeps the stored one.
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageRequest"
      responses:
        "200":
          description: Test result; `saved` says whether the settings were stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageTestResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/settings/storage/test:
    post:
      operationId: testStorageSettings
      summary: Run the live write/read/delete test without saving
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageRequest"
      responses:
        "200":
          description: Test result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageTestResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/settings/backup-key:
    get:
      operationId: getBackupKey
      summary: Whether the backup key exists and was confirmed
      tags: [settings]
      responses:
        "200":
          description: Key status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupKeyInfo"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: generateBackupKey
      summary: Generate the backup encryption key
      description: Returns the key for download. Refuses (409) if one exists, since replacing it would make every backup unreadable.
      tags: [settings]
      responses:
        "201":
          description: The new key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupKeyExport"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/orgs/{org}/leave:
    post:
      operationId: leaveOrg
      summary: Leave an organisation
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
      responses:
        "204":
          description: Left.
        default:
          $ref: "#/components/responses/Error"
...
components:
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginChallenge:
      type: object
      required: [challenge_id]
      properties:
        challenge_id:
          type: string
        enrollment:
          $ref: "#/components/schemas/LoginEnrollment"
...
    TotpRequest:
      type: object
      required: [challenge_id, code]
      properties:
        challenge_id:
          type: string
        code:
          type: string
    ReauthRequest:
      type: object
      required: [password, code]
      properties:
        password:
          type: string
          format: password
        code:
          type: string
...
    User:
      type: object
      required: [id, email, platform_role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        platform_role:
          type: string
          enum: [platform_admin, user]
...